game controller teleop code from Hello Robot
This Python 3 file from Hello Robot came from the GitHub repository found at https://github.com/hello-robot/stretch_body on September 1, 2021. This open source code was released with the Apache 2.0 license for use with the Stretch RE1 mobile manipulator. Students in the Robotic Caregivers class at Georgia Tech in the fall of 2021 modified this code and ran it on a Stretch RE1 in their third class on Wednesday, September 1, 2021.